+++ taste_speech/modules_taste/vq/vector_quantize_pytorch.py
@@ -471,7 +471,7 @@ class EuclideanCodebook(Module):
         needs_codebook_dim = x.ndim < 4
         sample_codebook_temp = default(sample_codebook_temp, self.sample_codebook_temp)
 
-        x = x.float()
+        x = x.to(self.embed.dtype)
 
         if needs_codebook_dim:
             x = rearrange(x, '... -> 1 ...')
